🔨 refactor(examples/with-redux): move redux to its own module and broadcast updates to modules that imports it
