+++ examples/with-redux/src/lib/redux/counterSlice.ts
@@ -1,12 +1,13 @@
-import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit'
+/* Core */
+import {
+  createAsyncThunk,
+  createSlice,
+  type PayloadAction,
+} from '@reduxjs/toolkit'
 
-import type { AppState, AppThunk } from '../../store'
-import { fetchCount } from './counterAPI'
-
-export interface CounterState {
-  value: number
-  status: 'idle' | 'loading' | 'failed'
-}
+/* Instruments */
+import { fetchCount } from '@/components'
+import type { AppState, AppThunk } from './store'
 
 const initialState: CounterState = {
   value: 0,
@@ -61,8 +62,6 @@ export const counterSlice = createSlice({
   },
 })
 
-export const { increment, decrement, incrementByAmount } = counterSlice.actions
-
 // The function below is called a selector and allows us to select a value from
 // the state. Selectors can also be defined inline where they're used instead of
 // in the slice file. For example: `useSelector((state: RootState) => state.counter.value)`
@@ -75,8 +74,12 @@ export const incrementIfOdd =
   (dispatch, getState) => {
     const currentValue = selectCount(getState())
     if (currentValue % 2 === 1) {
-      dispatch(incrementByAmount(amount))
+      dispatch(counterSlice.actions.incrementByAmount(amount))
     }
   }
 
-export default counterSlice.reducer
+/* Types */
+export interface CounterState {
+  value: number
+  status: 'idle' | 'loading' | 'failed'
+}
