+++ examples/with-redux/src/lib/redux/index.ts
@@ -1,0 +1,2 @@
+export * from './store'
+export * from './counterSlice'
